geom::Circle internal precision to double to address very large circles re: #2283

+++ src/cinder/GeomIo.cpp
@@ -1884,10 +1884,10 @@ void Circle::loadInto( Target *target, const AttribSet &/*requestedAttribs*/ ) c
 	normals.emplace_back( 0, 0, 1 );
 
 	// iterate the segments
-	const float tDelta = 1 / (float)mNumSubdivisions * 2.0f * 3.14159f;
-	float t = 0;
+	const double tDelta = 1 / (double)mNumSubdivisions * 2 * 3.14159;
+	double t = 0;
 	for( int s = 0; s <= mNumSubdivisions; s++ ) {
-		vec2 unit( math<float>::cos( t ), math<float>::sin( t ) );
+		vec2 unit( (float)math<double>::cos( t ), (float)math<double>::sin( t ) );
 		positions.emplace_back( mCenter + unit * mRadius );
 		texCoords.emplace_back( unit * 0.5f + vec2( 0.5f ) );
 		normals.emplace_back( 0, 0, 1 );
